Moved some quiz components/files around to see if yarn build works
