--- netlify.toml
@@ -1,4 +1,0 @@
-[build]
-  base = "/"
-  command = "yarn export"
-  publish = "out/"
